Navigate to a project dashboard and display budget info
Related to #176
