+++ frontend/contracts/MockToken.ts
@@ -43,3 +43,22 @@ export async function allowance(owner: string, spender: string, tokenAddress: st
   const contract = getMockTokenContract(signer, tokenAddress);
   return await contract.allowance(owner, spender);
 }
+
+// トークンのディテールを取得する関数
+export async function getTokenDetails(tokenAddress) {
+  const signer = getSigner();
+  const contract = getMockTokenContract(signer, tokenAddress);
+
+  try {
+    const decimals = await contract.decimals();
+    const symbol = await contract.symbol();
+    return { decimals, symbol };
+  } catch (error) {
+    console.error('Error fetching token details:', error);
+  }
+}
+
+// BigNumberを適切な数値に変換する関数
+export function formatTokenAmount(amount, decimals) {
+  return ethers.utils.formatUnits(amount, decimals);
+}
